# api.yaml
openapi: "3.1.0"
...
servers:
  - url: http://192.168.1.47:3000
    description: Local Server
  - url: http://localhost:3000
    description: Local Server
  - url: https://api.indhostel.com
    description: Production Server
  # - url: https://patiofy.smartaihr.com
  #   description: Production Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
info:
  title: Indhostels website Api
  description: API for indhostels website
  contact:
    name: support team
    url: http://indhostels.com/contactUs
    email: mukeshpabolu@techpixe.com
paths:
...
  /indhostels/auth/user/google/{googleId}:
    put:
      tags:
        - Google_OAuth2.0
      summary: Set a new password for a user
      description: Update the password for a user based on their ID.
      parameters:
        - name: googleId
          in: path
          required: true
          description: The ID of the user whose password is being updated.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newpassword
              properties:
                newpassword:
                  type: string
                  example: Ajay@110125
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: password updated successfully
        '400':
          description: Missing or invalid password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: password is required
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: object
  /indhostels/auth/user/signup:
    post:
      summary: User Signup using your details
      description: Creates a new user account. Sends a verification mail after successful signup.
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fullname
                - email
                - phone
                - password
                - confirmpassword
                - istermsandConditions
              properties:
                fullname:
                  type: string
                  example: Mukesh Pabolu
                email:
                  type: string
                  example: mukeshpabolu123@gmail.com
                phone:
                  type: number
                  example: 8374831031
                password:
                  type: string
                  example: Mukesh@123
                confirmpassword:
                  type: string
                  example: Mukesh@123
                istermsandConditions:
                  type: boolean
                  example: true
      responses:
        '200':
          description: User created and mail send to the email
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: mail send to your email, please verify
        '401':
          description: Email already taken
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: email already taken, try another
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: Some error occurred

  # /indhostels/auth/user/verifyaccount:
  #   post:
  #     summary: Verify user account using verifykey
  #     description: |
  #        This endpoint verifies the user account by verifing the verifykey sent to the user's email.
  #     tags:
  #       - users
  #     parameters:
  #       - name: verifykey
  #         in: query
  #         required: true
  #         schema:
  #           type: string
  #         description: The verifykey sent to the users email
  #     responses:
  #       "200":
  #         description: Account verified successfully.
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 message:
  #                   type: string
  #                   example: "Account verified successfully"
  #       "408":
  #         description: Verification time expired.
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 message:
  #                   type: string
  #                   example: "Time expired, please register"
  #       "500":
  #         description: Internal server error.
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 error:
  #                   type: string
  #                   example: "Internal Server Error"
  /indhostels/auth/user/verify:
    post:
      tags:
       - users
      summary: verify otp 
      description: vnerify u r otp to login or signup
      parameters:
        - name: phone 
          in: query 
          required: true 
          schema:
            type: number 
          description: The phone is required 
        - name: otp
          in: query
          required: true
          schema:
            type: number
          description: The otp sent to the users email 
      responses:
        "200":
          description: Account verified successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Account verified successfully"
        "408":
          description: Verification time expired.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Time expired, please register"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"
  /indhostels/auth/user/signin:
    post:
      tags:
        - users
      summary: Sign in a user
      description: Authenticates a user with phone number, and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - password
              properties:
                phone:
                  type: string
                  example: "8374831031"
                password:
                  type: string
                  example: "Test@123"
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: login successfully
                  JWTtoken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Authentication failed due to missing/invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: incorrect phone number or password does not match
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
  /indhostels/auth/user/me:
    get:
      summary: Get user by ID
      description: 
        This endpoint retrieves a user's details by their unique ID.
      tags:
        - users
      security:
        - BearerAuth: [] 
      responses:
        "200":
          description: User details retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    type: object
                    description: The user object containing details.
                    properties:
                      _id:
                        type: string
                        example: "65a2d5e2f1c3a12b34f6a789"
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "johndoe@example.com"
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "User not found"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Internal Server Error"
  /indhostels/auth/user/update:
    put:
      summary: Update User Profile
      description: Updates the logged-in user's fullname, phone and email.
      tags:
        - users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                  example: "John Doe"
                state:
                  type: string
                  example: "Hyderabad"
                city:
                  type: string
                  example: "Hyderabad"
                gender:
                  type: string 
                  example: male
                email:
                  type: string 
                  example: example@gmail.com
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Profile updated successfully
                  updatedUser:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: 609c0f58f3a9a2001510bdf1
                      fullname:
                        type: string
                        example: John Doe
                      email:
                        type: string
                        example: john.doe@example.com
                      phone:
                        type: number
                        example: 9000000000
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: Some error occurred
  /indhostels/auth/user/profilepic:
    post:
      summary: Upload or update user profile picture
      tags:
        - users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              image:
                - profilePic
              properties:
                image:
                  type: string
                  format: binary
                  description: User profile image file
      responses:
        '200':
          description: Profile picture uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  statuscode:
                    type: integer
                    example: 3
                  message:
                    type: string
                    example: User profile updated successfully
        '401':
          description: No file provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: profile image is required
                  error:
                    type: string
                    example: Bad Request
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 2
                  message:
                    type: string
                    example: User not Found
                  error:
                    type: string
                    example: Not Found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: error.message
  /indhostels/auth/user/password/forget:
     post:
      summary: Request password reset
      description: Sends a reset password link to the provided email address.
      tags:
        - users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: marojuajay111@gmail.com
      responses:
        '200':
          description: Password reset link sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: reset password link sent to the email
        '401':
          description: Incorrect email or validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: incorrect email
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: Detailed error message
  /indhostels/auth/user/password/setNew:
    put:
      summary: Set a new password for a user
      tags:
        - users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: mukeshpabolu123@gmail.com
                password:
                  type: string
                  format: password
                  example: mukesh@110125

      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: new password updated , please login
        '400':
          description: Missing phone or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: All fields are required
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal server Error
                  error:
                    type: string
                    example: Error message details
  /indhostels/auth/user/password/change:
    post:
      summary: Reset user password
      description: |
        This endpoint allows a user to reset their password by providing their old password and a new password. 
        If the old password matches, the new password is updated, and the user is logged out.
      tags:
        - users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldpassword:
                  type: string
                  description: The user's current password.
                newpassword:
                  type: string
                  description: The new password to set.
                confirmpassword:
                  type: string
                  description: The confirm password to set.
      responses:
        "200":
          description: Password updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "password updated successfully, Please Login"
        "400":
          description: Missing required fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "all fields are required"
        "401":
          description: Old password does not match.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "password does not match"
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"
  /indhostels/auth/user/logout:
    put:
      summary: Sign out user
      description: |
        This endpoint logs out the user by clearing the authentication token cookie. 
        The cookie is configured with `httpOnly`, `secure`, and `sameSite` attributes for security.
      tags:
        - users
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User logged out successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "user logged out successfully"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"
  /indhostels/auth/user/deactivateaccount:
    put:
      summary: Deactivate the user account
      description: |
        This endpoint deactivates the user account by updating the status to "deactivated".
      tags:
        - users
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User deactivated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User deactivated successfully"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"
 
 
  # /indhostels/auth/user/deleteaccount:
  #   delete:
  #     summary: Delete The User Account
  #     description: |
  #       This endpoint deletes the user and all related data.
  #     tags:
  #       - users
  #     security:
  #       - BearerAuth: []
  #     responses:
  #       "200":
  #         description: User deleted successfully.
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 success:
  #                   type: boolean
  #                   example: true
  #                 message:
  #                   type: string
  #                   example: "User deleted successfully"
  #       "500":
  #         description: Internal server error.
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 error:
  #                   type: string
  #                   example: "Internal Server Error"
  
  
#✅ user product calls ✅

  #Products Routes
...
  /indhostels/auth/user/query:
    post:
      summary: Submit a Contact Us Message
      description: Allows an authenticated user to submit a message via the contact form.
      tags:
        - contactus
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                  example: "Pabolu Mukesh Narayana"
                email:
                  type: string
                  example: "mukeshpabolu123@gmail.com"
                phone:
                  type: string
                  example: "8374831031"
                subject:
                  type: string
                  example: "I have an issue with my booking."
                message:
                  type: string
                  example: "I have an issue with my booking."
      responses:
        200:
          description: Successfully submitted the contact message.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Submitted successfully"
        400:
          description: Bad request, missing or invalid input.
        401:
          description: Unauthorized, user not logged in.
        500:
          description: Internal Server Error.

#✅ user order calls ✅
...
  /indhostels/auth/user/wishlist:
    post:
      summary: Add accommodation to wishlist
      description: Add accommodation to wishlist
      tags:
        - user-wishlist 
      security:
        - BearerAuth: [] 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accommodationid:
                  type: string
      responses:
        '200':
          description: Wishlist added successfully 
        '404':
          description: Wishlist not found 
        '500':
          description: Internal Server Error 
  /indhostels/auth/user/getwishlist:
    get:
      summary: Get wishlists of perticular user
      description: Get wishlists of perticular user
      tags:
        - user-wishlist 
      security:
        - BearerAuth: [] 
      responses:
        '200':
          description: Wishlist retrived successfully 
        '404':
          description: Wishlist not found 
        '500':
          description: Internal Server Error 
  /indhostels/auth/user/deletewishlist:
    delete:
      summary: Delete wishlist of perticular user
      description: Delete wishlist of perticular user
      tags:
        - user-wishlist 
      security:
        - BearerAuth: [] 
      parameters:
        - in: query
          name: wishlistid
          required: true
          schema:
            type: string
          description: Wishlist ID (MongoDB ObjectId)
      responses:
        '200':
          description: Wishlist deleted successfully 
        '404':
          description: Wishlist not found 
        '500':
          description: Internal Server Error 


# ✅ ratings ✅
...
  /indhostels/auth/user/notification:
    get:
      summary: Get all notifications of perticular user
      description: Get all notifications of perticular user
      tags:
        - user-notifications 
      security:
        - BearerAuth: [] 
      responses:
        '200':
          description: Notification retrived successfully 
        '404':
          description: Notification not found 
        '500':
          description: Internal Server Error
  /indhostels/auth/user/notification/{notificationid}:
    get:
      summary: Get perticular notification using the id
      tags: 
       - user-notifications
      security:
        - BearerAuth: [] 
      parameters:
        - in: path
          name: notificationid
          required: true
          schema:
            type: string
            description: Notification ID (MongoDB ObjectId)
      responses:
        '200':
          description: Notification retrived successfully 
        '404':
          description: Notification not found 
        '500':
          description: Internal Server Error

#✅ Help and Support ✅ 
...